+++ example/test/calculator/calculator.bdd_scenarios.dart
@@ -12,7 +12,6 @@ class AddTwoNumbersScenario {
   static Future<void> theResultShouldBe3(WidgetTester tester) async {
     // TODO: Implement Then the result should be 3
   }
-
 }
 
 class SubtractTwoNumbersScenario {
@@ -27,7 +26,6 @@ class SubtractTwoNumbersScenario {
   static Future<void> theResultShouldBe2(WidgetTester tester) async {
     // TODO: Implement Then the result should be 2
   }
-
 }
 
 class MultiplyTwoNumbersScenario {
@@ -42,7 +40,6 @@ class MultiplyTwoNumbersScenario {
   static Future<void> theResultShouldBe6(WidgetTester tester) async {
     // TODO: Implement Then the result should be 6
   }
-
 }
 
 class DivideTwoNumbersScenario {
@@ -57,6 +54,4 @@ class DivideTwoNumbersScenario {
   static Future<void> theResultShouldBe(WidgetTester tester, String result) async {
     // TODO: Implement Then the result should be <result>
   }
-
 }
-