+++ lib/src/extensions/string_x.dart
@@ -14,10 +14,7 @@ extension StringX on String {
   }
 
   String get name {
-    return split(' ')
-        .where((word) => word.isNotEmpty)
-        .map((word) => word[0].toUpperCase() + word.substring(1).toLowerCase())
-        .join('');
+    return split(' ').where((word) => word.isNotEmpty).map((word) => word[0].toUpperCase() + word.substring(1).toLowerCase()).join('');
   }
 
   String get toScenarioClassName {
@@ -25,9 +22,7 @@ extension StringX on String {
   }
 
   String get snakeCaseToCamelCase {
-    return split('_')
-        .map((word) => word[0].toLowerCase() + word.substring(1).toLowerCase())
-        .join('');
+    return split('_').map((word) => word[0].toLowerCase() + word.substring(1).toLowerCase()).join('');
   }
 
   String get toSnakeCase {