+++ lib/src/feature/builder/bdd_builders/scenario_file_builder.dart
@@ -7,7 +7,7 @@ import '../../../extensions/string_x.dart';
 class ScenariosFileBuilder {
   Future<void> build(BuildStep buildStep, Feature feature) async {
     final inputId = buildStep.inputId;
-    final scenarioOutputId = inputId.changeExtension('.bdd_scenarios.dart');
+    final scenarioOutputId = inputId.changeExtension('.bdd_scenarios.g.dart');
 
     final scenarioContent = await buildScenarioFile(feature);
     await buildStep.writeAsString(scenarioOutputId, scenarioContent);